Move drop handler out of Cell.js
