+++ frontend/components/Cell.js
@@ -4,6 +4,7 @@ import { CellOutput } from "./CellOutput.js"
 import { CellInput } from "./CellInput.js"
 import { RunArea, useMillisSinceTruthy } from "./RunArea.js"
 import { cl } from "../common/ClassTable.js"
+import { useDropHandler } from "./dropHandler.js"
 
 /**
  * @typedef {Object} CodeState
@@ -92,9 +93,7 @@ export const Cell = ({
 }) => {
     // cm_forced_focus is null, except when a line needs to be highlighted because it is part of a stack trace
     const [cm_forced_focus, set_cm_forced_focus] = useState(null)
-    const [dragActive, setDragActiveFast] = useState(false)
-    const [savingFile, setSavingFile] = useState(false)
-    const setDragActive = useMemo(() => _.debounce(setDragActiveFast, 200), [setDragActiveFast])
+    const { savingFile, dragActive, eventHandler } = useDropHandler(requests, on_change, cell_id)
     const localTimeRunning = 10e5 * useMillisSinceTruthy(running)
     useEffect(() => {
         const focusListener = (e) => {
@@ -120,73 +119,7 @@ export const Cell = ({
     const class_code_folded = code_folded && cm_forced_focus == null
 
     let show_input = errored || class_code_differs || !class_code_folded
-    const prepareFileBase64 = (file) =>
-        new Promise((resolve, reject) => {
-            const { name, type } = file
-            const fr = new FileReader()
-            fr.onerror = () => reject("Failed to read file!")
-            fr.onloadstart = () => {}
-            fr.onprogress = ({ loaded, total }) => {}
-            fr.onload = () => {}
-            fr.onloadend = ({ target: { result } }) => resolve({ fileBase64: String(result).replace(/.*base64,/, ""), name, type })
-            fr.readAsDataURL(file)
-        })
 
-    const uploadAndCreateCodeTemplate = async (file) => {
-        if (!(file instanceof File)) return " #  File can't be read"
-        setSavingFile(true)
-
-        const {
-            message: { success, code },
-        } = await prepareFileBase64(file).then(
-            (preparedObj) => {
-                return requests.write_file(cell_id, preparedObj)
-            },
-            () => alert("Pluto can't save this file 😥")
-        )
-        setSavingFile(false)
-        if (!success) {
-            alert("Pluto can't save this file 😥")
-            return "# File save failed"
-        }
-        if (code) return code
-        alert("Pluto doesn't know what to do with this file 😥. Feel that's wrong? Open an issue!")
-        return ""
-    }
-    const eventHandler = (ev) => {
-        // dataTransfer is in Protected Mode here. see type, let Pluto DropRuler handle it.
-        if (ev.dataTransfer.types[0] === "text/pluto-cell") return
-        switch (ev.type) {
-            case "cmdrop":
-            case "drop":
-                console.log("runs?")
-                ev.preventDefault() // don't file open
-                setDragActive(false)
-                if (!ev.dataTransfer.files.length) {
-                    return
-                }
-                uploadAndCreateCodeTemplate(ev.dataTransfer.files[0]).then((code) => {
-                    if (code) {
-                        on_change(code)
-                        requests.change_remote_cell(cell_id, code)
-                    }
-                })
-                break
-            case "dragover":
-                ev.preventDefault()
-                ev.dataTransfer.dropEffect = "copy"
-                setDragActive(true)
-                setTimeout(() => setDragActive(false), 500)
-                break
-            case "dragenter":
-                setDragActiveFast(true)
-                break
-            case "dragleave":
-                setDragActive(false)
-                break
-            default:
-        }
-    }
     return html`
         <pluto-cell
             onDragOver=${eventHandler}
